simple-mcp-server/samples/sample-maf-agent/pyproject.toml
[project]
name = "neo4j-provider-samples"
version = "0.1.0"
description = "Sample applications for agent-framework-neo4j and MCP server integration"
requires-python = ">=3.10"
dependencies = [
    "agent-framework-neo4j[azure]",
    "agent-framework-azure-ai>=1.0.0b",
    "agent-framework>=1.0.0b",  # Required for MCPStreamableHTTPTool (not re-exported by azure-ai)
    "httpx>=0.27.0",  # Required for authenticated HTTP client with MCP tools
    "python-dotenv",
]

[project.scripts]
start-samples = "samples.shared.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
prerelease = "allow"
